% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeval.R
\name{modeval}
\alias{modeval}
\title{Evaluate model using measured values}
\usage{
modeval(meas, mod)
}
\arguments{
\item{meas}{Measured values.}

\item{mod}{modelled values.}
}
\value{
Evaluation of \code{mod} using \code{meas}.
}
\description{
Evaluate model using measured values
}
